Create Encrypter.java
this class is responsible for reading input text, creating a frequency table based on that text, then constructing a tree (also the codex) to map characters to binary patterns based on their frequencies and the path you must take in the tree to get to that character. It then creates an output file containing the codebook for the characters used in the input file, followed by a message written in a binary pattern.